+++ changelog.md
@@ -1,3 +1,13 @@
+0.1.0 (2018-09-11)
+==================
+
+* Dropped `ClientError` and `ServerError` interfaces, they aren't useful in
+  TS.
+* Added `HttpError` and `HttpProblem` interfaces.
+* Added `isHttpError` type guard.
+* Added `isClientError` and `isServerError` helpers.
+
+
 0.0.1 (2018-09-11)
 ==================